api/pyproject.toml
[build-system]
requires = ["flit_core >= 3.9"]
build-backend = "flit_core.buildapi"

[project]
name = "verse"
version = "0.1.0"
description = "An API for a RAG-enhanced model designed for study in the Yale OpenCourseware English literature program."
authors = [{name = "Natasha Ravinand", email = "natasharavinand@gmail.com"}]
dependencies = [
    "aiohttp==3.9.3",
    "aiosignal==1.3.1",
    "altair==5.2.0",
    "annotated-types==0.6.0",
    "anyio==4.3.0",
    "asgiref==3.7.2",
    "attrs==23.2.0",
    "backoff==2.2.1",
    "bcrypt==4.1.2",
    "beautifulsoup4==4.12.3",
    "blinker==1.7.0",
    "build==1.1.1",
    "cachetools==5.3.3",
    "certifi==2024.2.2",
    "chardet==5.2.0",
    "charset-normalizer==3.3.2",
    "chroma-hnswlib==0.7.3",
    "chromadb==0.4.24",
    "click==8.1.7",
    "coloredlogs==15.0.1",
    "coverage==7.4.4",
    "dataclasses-json==0.6.4",
    "dataclasses-json-speakeasy==0.5.11",
    "Deprecated==1.2.14",
    "distro==1.9.0",
    "emoji==2.10.1",
    "fastapi==0.110.0",
    "filelock==3.13.1",
    "filetype==1.2.0",
    "Flask==3.0.2",
    "Flask-Cors==4.0.0",
    "flatbuffers==24.3.7",
    "frozenlist==1.4.1",
    "fsspec==2024.2.0",
    "gitdb==4.0.11",
    "GitPython==3.1.42",
    "google-auth==2.28.2",
    "googleapis-common-protos==1.63.0",
    "grpcio==1.62.1",
    "h11==0.14.0",
    "httpcore==1.0.4",
    "httptools==0.6.1",
    "httpx==0.27.0",
    "huggingface-hub==0.21.4",
    "humanfriendly==10.0",
    "idna==3.6",
    "importlib-metadata==6.11.0",
    "importlib_resources==6.3.0",
    "iniconfig==2.0.0",
    "itsdangerous==2.1.2",
    "Jinja2==3.1.3",
    "joblib==1.3.2",
    "jsonpatch==1.33",
    "jsonpath-python==1.0.6",
    "jsonpointer==2.4",
    "jsonschema==4.21.1",
    "jsonschema-specifications==2023.12.1",
    "kubernetes==29.0.0",
    "langchain==0.1.12",
    "langchain-community==0.0.28",
    "langchain-core==0.1.31",
    "langchain-openai==0.0.8",
    "langchain-text-splitters==0.0.1",
    "langdetect==1.0.9",
    "langsmith==0.1.24",
    "lxml==5.1.0",
    "markdown-it-py==3.0.0",
    "MarkupSafe==2.1.5",
    "marshmallow==3.20.2",
    "mdurl==0.1.2",
    "mmh3==4.1.0",
    "monotonic==1.6",
    "mpmath==1.3.0",
    "multidict==6.0.5",
    "mypy-extensions==1.0.0",
    "nltk==3.8.1",
    "numpy==1.26.4",
    "oauthlib==3.2.2",
    "onnxruntime==1.17.1",
    "openai==1.14.0",
    "opentelemetry-api==1.23.0",
    "opentelemetry-exporter-otlp-proto-common==1.23.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.23.0",
    "opentelemetry-instrumentation==0.44b0",
    "opentelemetry-instrumentation-asgi==0.44b0",
    "opentelemetry-instrumentation-fastapi==0.44b0",
    "opentelemetry-proto==1.23.0",
    "opentelemetry-sdk==1.23.0",
    "opentelemetry-semantic-conventions==0.44b0",
    "opentelemetry-util-http==0.44b0",
    "orjson==3.9.15",
    "overrides==7.7.0",
    "packaging==23.2",
    "pandas==2.2.1",
    "pillow==10.2.0",
    "pluggy==1.4.0",
    "posthog==3.5.0",
    "protobuf==4.25.3",
    "pulsar-client==3.4.0",
    "pyarrow==15.0.1",
    "pyasn1==0.5.1",
    "pyasn1-modules==0.3.0",
    "pydantic==2.6.4",
    "pydantic_core==2.16.3",
    "pydeck==0.8.1b0",
    "Pygments==2.17.2",
    "PyPika==0.48.9",
    "pyproject_hooks==1.0.0",
    "pytest==8.1.1",
    "pytest-cov==5.0.0",
    "pytest-mock==3.14.0",
    "python-dateutil==2.8.2",
    "python-dotenv==1.0.1",
    "python-iso639==2024.2.7",
    "python-magic==0.4.27",
    "pytz==2024.1",
    "PyYAML==6.0.1",
    "rapidfuzz==3.6.1",
    "referencing==0.33.0",
    "regex==2023.12.25",
    "requests==2.31.0",
    "requests-oauthlib==1.4.0",
    "rich==13.7.1",
    "rpds-py==0.18.0",
    "rsa==4.9",
    "six==1.16.0",
    "smmap==5.0.1",
    "sniffio==1.3.1",
    "soupsieve==2.5",
    "SQLAlchemy==2.0.28",
    "starlette==0.36.3",
    "streamlit==1.32.1",
    "sympy==1.12",
    "tabulate==0.9.0",
    "tenacity==8.2.3",
    "tiktoken==0.6.0",
    "tokenizers==0.15.2",
    "toml==0.10.2",
    "toolz==0.12.1",
    "tornado==6.4",
    "tqdm==4.66.2",
    "typer==0.9.0",
    "typing-inspect==0.9.0",
    "typing_extensions==4.9.0",
    "tzdata==2024.1",
    "unstructured==0.12.6",
    "unstructured-client==0.18.0",
    "urllib3==1.26.18",
    "uvicorn==0.28.0",
    "uvloop==0.19.0",
    "watchdog==4.0.0",
    "watchfiles==0.21.0",
    "websocket-client==1.7.0",
    "websockets==12.0",
    "Werkzeug==3.0.1",
    "wrapt==1.16.0",
    "yarl==1.9.4",
    "zipp==3.18.0"
]
readme = "README.md"